Moved SetAllServicesPublicPass to eZ\Publish\API\Repository\Tests\Container\Compiler
