+++ eZ/Publish/SPI/Tests/FieldType/FileBaseIntegrationTest.php
@@ -2,7 +2,7 @@
 
 namespace eZ\Publish\SPI\Tests\FieldType;
 
-use eZ\Publish\Core\Base\Container\Compiler\SetAllServicesPublicPass;
+use eZ\Publish\API\Repository\Tests\Container\Compiler\SetAllServicesPublicPass;
 use eZ\Publish\Core\IO\IOServiceInterface;
 use RecursiveIteratorIterator;
 use RecursiveDirectoryIterator;
